backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.10.0
websockets==15.0.1
pillow==10.1.0
google-cloud-vision==3.4.2
google-generativeai==0.3.0
requests==2.31.0
psycopg2-binary==2.9.9
